+++ tests/ServiceProviderTest.php
@@ -104,6 +104,71 @@ class AliasServiceProvider extends ServiceProvider
     }
 }
 
+class TestDependency
+{
+
+    public function getValue()
+    {
+        return 'test';
+    }
+}
+
+class TestProviderInjection extends ServiceProvider
+{
+
+    /**
+     * @var TestDependency
+     */
+    protected $dependency;
+
+    public function __construct(Container $container, TestDependency $dependency)
+    {
+        parent::__construct($container);
+
+        $this->dependency = $dependency;
+    }
+
+    public function register()
+    {
+        // TODO: Implement register() method.
+    }
+
+    public function getContainer() {
+        return $this->container;
+    }
+
+    public function getDependency()
+    {
+        return $this->dependency;
+    }
+}
+
+class TestProviderPrimitiveInjection extends ServiceProvider
+{
+
+    /**
+     * @var string
+     */
+    protected $value;
+
+    public function __construct(Container $container, $value)
+    {
+        parent::__construct($container);
+
+        $this->value = $value;
+    }
+
+    public function register()
+    {
+        // TODO: Implement register() method.
+    }
+
+    public function getValue()
+    {
+        return $this->value;
+    }
+}
+
 class ServiceProviderTest extends TestCase
 {
     /**
@@ -292,5 +357,44 @@ class ServiceProviderTest extends TestCase
             $container->getProvider(AliasServiceProvider::class),
             $container->getProvider('other-service-provider')
         );
+    }
+
+    /**
+     * Test we can create service providers with additional concrete dependencies.
+     *
+     * @test
+     */
+    public function should_automatically_inject_concrete_instances_into_extended_providers()
+    {
+        $container = new Container();
+
+        $container->register(TestProviderInjection::class);
+        $this->assertTrue($container->has(TestProviderInjection::class));
+        $this->assertTrue($container->has(TestDependency::class));
+
+        $provider = $container->get(TestProviderInjection::class);
+
+        $this->assertSame($container, $provider->getContainer());
+        $this->assertInstanceOf(TestDependency::class, $provider->getDependency());
+        $this->assertSame('test', $provider->getDependency()->getValue());
+    }
+
+    /**
+     * Test we can configure the container with providers that have primitives
+     * in their constructor.
+     *
+     * @test
+     */
+    public function should_allow_binding_providers_with_primitive_values()
+    {
+        $container = new Container();
+
+        $container->when(TestProviderPrimitiveInjection::class)
+            ->needs('$value')
+            ->give('test');
+
+        $container->register(TestProviderPrimitiveInjection::class);
+        $this->assertTrue($container->has(TestProviderPrimitiveInjection::class));
+        $this->assertSame('test', $container->get(TestProviderPrimitiveInjection::class)->getValue());
     }
 }